+++ CMakeLists.txt
@@ -421,13 +421,8 @@ add_subdirectory(./metautils/lib)
 add_subdirectory(./cache)
 add_subdirectory(./cluster/lib)
 
-add_subdirectory(./cluster/conscience)
 add_subdirectory(./cluster/module)
 add_subdirectory(./cluster/tools)
-add_subdirectory(./gridd/main)
-add_subdirectory(./gridd/plugins/msg_fallback)
-add_subdirectory(./gridd/plugins/msg_ping)
-add_subdirectory(./gridd/plugins/msg_stats)
 add_subdirectory(./meta0v2)
 add_subdirectory(./meta1v2)
 add_subdirectory(./meta2v2)
