--- cluster/gridstorage.conf
@@ -1,5 +1,0 @@
-## Configuration example
-#
-#[NS_TEST]
-#conscience=127.0.0.1:6000
-#
