--- cluster/module/module.h
@@ -1,75 +1,0 @@
-/*
-OpenIO SDS cluster conscience
-Copyright (C) 2014 Worldline, as part of Redcurrant
-Copyright (C) 2015-2017 OpenIO SAS, as part of OpenIO SDS
-
-This program is free software: you can redistribute it and/or modify
-it under the terms of the GNU Affero General Public License as
-published by the Free Software Foundation, either version 3 of the
-License, or (at your option) any later version.
-
-This program is distributed in the hope that it will be useful,
-but WITHOUT ANY WARRANTY; without even the implied warranty of
-MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-GNU Affero General Public License for more details.
-
-You should have received a copy of the GNU Affero General Public License
-along with this program.  If not, see <http://www.gnu.org/licenses/>.
-*/
-
-#ifndef OIO_SDS__cluster__module__module_h
-# define OIO_SDS__cluster__module__module_h 1
-
-#if 0
-// FIXME dead code
-# define NAME_MSGNAME_CS_GETNS  "REQ_CS_GETNS"
-#endif
-
-# define NAME_MSGNAME_CS_GET_NSINFO   "CS_CFG"
-# define NAME_MSGNAME_CS_GET_SRVNAMES "CS_TYP"
-# define NAME_MSGNAME_CS_GET_SRV      "CS_SRV"
-# define NAME_MSGNAME_CS_PUSH_SRV     "CS_PSH"
-# define NAME_MSGNAME_CS_RM_SRV       "CS_DEL"
-
-# define NAME_GROUPNAME_STORAGE_POLICY   "STORAGE_POLICY"
-# define NAME_GROUPNAME_DATA_SECURITY    "DATA_SECURITY"
-# define NAME_GROUPNAME_DATA_TREATMENTS  "DATA_TREATMENTS"
-# define NAME_GROUPNAME_SERVICE_POOLS    "pools"
-# define NAME_GROUPNAME_SCORE_EXPR       "score_expr"
-# define NAME_GROUPNAME_SCORE_TIMEOUT    "score_timeout"
-# define NAME_GROUPNAME_SCORE_VARBOUND   "score_variation_bound"
-# define NAME_GROUPNAME_SCORE_LOCK       "score_lock_at_first_register"
-# define STG_CONF_STR_MAX_LENGTH			50
-
-# define KEY_NAMESPACE "namespace"
-# define KEY_ALERT_LIMIT "alert_frequency_limit"
-# define KEY_SCORE_TIMEOUT "score_timeout"
-# define KEY_SCORE_EXPR "score_expr"
-# define KEY_STG_CONF "storage_conf"
-# define KEY_SVC_CONF "service_conf"
-# define KEY_SCORE_VARBOUND "score_variation_bound"
-# define KEY_SCORE_LOCK "lock_at_first_register"
-
-# define KEY_POOL_TARGETS "targets"
-# define KEY_POOL_MASK    "mask"
-
-# define GROUP_PREFIX_TYPE "type:"
-# define GROUP_PREFIX_POOL "pool:"
-
-# define KEY_SERIALIZE_SRVINFO_TAGS   "serialize_srvinfo_tags"
-# define DEF_SERIALIZE_SRVINFO_TAGS   TRUE
-
-# define KEY_SERIALIZE_SRVINFO_STATS  "serialize_srvinfo_stats"
-# define DEF_SERIALIZE_SRVINFO_STATS  FALSE
-
-# define TIME_DEFAULT_ALERT_LIMIT 300L
-
-# define EXPR_DEFAULT_META0 "100"
-# define EXPR_DEFAULT_META1 "root(2,((num stat.cpu)*(num stat.io)))"
-# define EXPR_DEFAULT_META2 "root(2,((num stat.cpu)*(num stat.io)))"
-# define EXPR_DEFAULT_RAWX  "root(3,((num stat.cpu)*(num stat.io)*(num stat.space)))"
-
-# define NB_BROKEN_LINES 64U
-# define NB_SRV_ELEMENTS 128U
-
-#endif /*OIO_SDS__cluster__module__module_h*/
