--- gridd/plugins/msg_fallback/CMakeLists.txt
@@ -1,13 +1,0 @@
-add_definitions(-DG_LOG_DOMAIN="oio.gridd.fallback")
-set(CMAKE_C_FLAGS "${CMAKE_C_FLAGS} -Werror")
-
-include_directories(BEFORE . ../.. ../../..)
-
-add_library(msg_fallback MODULE msg_fallback.c)
-set_target_properties(msg_fallback PROPERTIES PREFIX "" SUFFIX .so)
-target_link_libraries(msg_fallback metautils ${GLIB2_LIBRARIES})
-
-install(TARGETS msg_fallback
-        LIBRARY DESTINATION ${GRIDD_PLUGINS_DIRECTORY}
-        RUNTIME DESTINATION bin)
-
