--- gridd/plugins/msg_stats/msg_stats.h
@@ -1,30 +1,0 @@
-/*
-OpenIO SDS gridd
-Copyright (C) 2014 Worldline, as part of Redcurrant
-Copyright (C) 2015-2017 OpenIO SAS, as part of OpenIO SDS
-
-This program is free software: you can redistribute it and/or modify
-it under the terms of the GNU Affero General Public License as
-published by the Free Software Foundation, either version 3 of the
-License, or (at your option) any later version.
-
-This program is distributed in the hope that it will be useful,
-but WITHOUT ANY WARRANTY; without even the implied warranty of
-MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-GNU Affero General Public License for more details.
-
-You should have received a copy of the GNU Affero General Public License
-along with this program.  If not, see <http://www.gnu.org/licenses/>.
-*/
-
-#ifndef OIO_SDS__gridd__plugins__msg_stats__msg_stats_h
-# define OIO_SDS__gridd__plugins__msg_stats__msg_stats_h 1
-
-#define MSGKEY_PATTERN "PATTERN"
-
-#define MSGFIELD_STATPREFIX "stat:"
-
-#define MSGFIELD_STATKEY_MAXLEN 256
-#define MSGFIELD_STATVAL_MAXLEN 64
-
-#endif /*OIO_SDS__gridd__plugins__msg_stats__msg_stats_h*/
