+++ metautils/lib/metautils.h
@@ -56,7 +56,6 @@ License along with this library.
 # include <metautils/lib/storage_policy.h>
 # include <metautils/lib/common_main.h>
 # include <metautils/lib/volume_lock.h>
-# include <metautils/lib/expr.h>
 # include <metautils/lib/rrd.h>
 # include <metautils/lib/lb.h>
